+++ .gitignore
@@ -1,1 +1,5 @@
 .vs
+/bin
+/obj
+*.sln
+cafeRecAPI.sln
