+++ Models/CafeLocation.cs
@@ -6,7 +6,6 @@
         {
             public int CafeId { get; set; }
             public virtual Cafe Cafe { get; set; }
-
             public int SearchLocationId { get; set; }
             public virtual SearchLocation SearchLocation { get; set; }
         }
