+++ Models/Review.cs
@@ -1,9 +1,10 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Text.Json.Serialization;
 
 namespace cafeRecAPI.Models
 {
-    public class Tag
+    public class Review
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
@@ -11,6 +12,10 @@ namespace cafeRecAPI.Models
         public int Id { get; set; }
 
         [Required]
-        public string? TagName { get; set; }
+        public string? ReviewText { get; set; }
+        public int? CafeId { get; set; }
+
+        [JsonIgnore]
+        public virtual Cafe? Cafe { get; set; }
     }
 }
