+++ cafeRecAPI.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34525.116
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "cafeRecAPI", "cafeRecAPI\cafeRecAPI.csproj", "{75FF3FD8-DFC3-4420-A706-B02C8E9748A2}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "cafeRecAPI", "cafeRecAPI.csproj", "{75FF3FD8-DFC3-4420-A706-B02C8E9748A2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
