--- cafeRecAPI/Dockerfile
@@ -1,30 +1,0 @@
-#See https://aka.ms/customizecontainer to learn how to customize your debug container and how Visual Studio uses this Dockerfile to build your images for faster debugging.
-
-FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
-USER app
-WORKDIR /app
-EXPOSE 8080
-EXPOSE 8081
-
-FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
-ARG BUILD_CONFIGURATION=Release
-WORKDIR /src
-COPY ["cafeRecAPI/cafeRecAPI.csproj", "cafeRecAPI/"]
-RUN dotnet restore "./cafeRecAPI/./cafeRecAPI.csproj"
-COPY . .
-WORKDIR "/src/cafeRecAPI"
-RUN dotnet build "./cafeRecAPI.csproj" -c $BUILD_CONFIGURATION -o /app/build
-
-FROM build AS publish
-ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "./cafeRecAPI.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app/publish .
-
-COPY ["Service/*.so", "./Service/"]  # Copy shared library files
-ENV LD_LIBRARY_PATH=./Service:$LD_LIBRARY_PATH
-
-
-ENTRYPOINT ["dotnet", "cafeRecAPI.dll"]
